lilypond: handle case where `\\` does not appear within a polyphony
It may happen that a string contains `\\`, which will wrongly be caught as starting a new polyphonic voice. If we see that no polyphony is currently active, we simply move onward to avoid indexing a nil value.

This likely means that any `\\` are also going to be wrongly caught _inside_ a polyphonic passage. The lack of a syntax parser makes handling such cases more tricky and less general.