fix(conform.nvim): Chose between multiple formatters
It appears as though a previous commit removed the ability to specify
multiple formatters, and have conform pick the first available one from
a list.

This fix reimplements this functionality and allows specifiying multiple
formatters to be run in sequence, while also allowing to pick between
sets of formatters at each step.