fix(midi-input): Use GitHub repo by default instead of local folder
Using this configuration on a different system, one can of course not
expect the folder in question to be present. This always leads to an
annoying error message about the plugin being unable to load.

This fix simply opts towards using the GitHub repo directly. When
development on the plugin is necessary, one may temporarily set the
`dir` option again.