chore(nix): Prepare commands to refactor nix code more easily
Nix allows nesting attributes or chaining them using the `.`. However,
if the dot syntax leads to repetition of the same attribute in the same
level, then we should ideally merge the repeated attributes
sub-attributes.

At the same time, it would also be useful to un-merge such attributes in
case we want to move things around or edit them more easily.