feat(nix): Implement merging of attribute under cursor
Currently very rudimentary. It does not handle the case where the cursor
is on a nested attribute, i.e. `x.y.z` and you have your cursor on `y`
or `z`.

It currently seems to be fetching the right values from other
attributes, but does not compose it properly.

Furthermore, it does not at all handle the case where the value of the
first attribute is not an attribute set.