fix(lsp): Replace "require('lspconfig')" for vim.lsp.config
The require is deprecated and can no longer be used in any way shape or
form. Originally, we retrieve the lspconfig provided default value for
the filetypes so that we can properly extend (not overwrite) it with our
own.

To still achieve this, we instead read the `lsp/ltex.lua` from the
runtime path files and filter the files so that we only retain the one
provided by the lspconfig plugin. This way, we can still obtain the
default definition from the plugin.