fix(instant.nvim): Saving joined buffers breaks the syncing
Note that this patch cannot discern between client and server. Ideally,
on the server we would like to keep saving the buffers. However, on the
client, saving the buffer may lead to issues in that the syncing breaks.

As a slight inconvenience for the server, but to prevent syncing issues
with the client, both will have their buffers in readonly.